+++ scripts/macros/2014/monsterFeatures/generic/failedByAmount.js
@@ -1,0 +1,48 @@
+import {itemUtils, workflowUtils} from '../../../../utils.js';
+async function use({workflow}) {
+    let config = itemUtils.getGenericFeatureConfig(workflow.item, 'failedByAmount');
+    let activities = config.activities;
+    if (activities?.length && !activities.includes(workflow.activity.id)) return;
+    let targets = workflow.failedSaves.filter(i => workflow.tokenSaves[i.document.uuid].total <= (workflow.saveDC - config.amount));
+    let triggerActivities = config.triggerActivities.map(i => workflow.item.system.activities.find(j => j.id === i));
+    await Promise.all(targets.map(async i => {
+        for (let j of triggerActivities) {
+            await workflowUtils.syntheticActivityRoll(j, [i]);
+        }
+    }));
+}
+export let failedByAmount = {
+    name: 'failedByAmount',
+    translation: 'CHRISPREMADES.Macros.FailedByAmount.Name',
+    version: '1.3.39',
+    midi: {
+        item: [
+            {
+                pass: 'rollFinished',
+                macro: use,
+                priority: 50
+            }
+        ]
+    },
+    isGenericFeature: true,
+    genericConfig: [
+        {
+            value: 'activities',
+            label: 'CHRISPREMADES.Config.Activities',
+            type: 'activities',
+            default: []
+        },
+        {
+            value: 'triggerActivities',
+            label: 'CHRISPREMADES.Macros.ActivityOnRest.TriggerActivities',
+            type: 'activities',
+            default: []
+        },
+        {
+            value: 'amount',
+            label: 'CHRISPREMADES.Macros.FailedByAmount.Amount',
+            type: 'number',
+            default: '5'
+        }
+    ]
+};